Organizations modularization into views, components and services
