--- frontend/src/modules/Misc/views/Organizations/index.js
@@ -1,5 +1,0 @@
-import OrganizationList from './OrganizationList';
-import OrganizationListItem from './OrganizationListItem';
-import OrganizationView from './OrganizationView';
-
-export { OrganizationList, OrganizationListItem, OrganizationView };
