+++ frontend/src/modules/Organizations/components/OrganizationListItem.js
@@ -15,9 +15,9 @@ import * as FaIcons from 'react-icons/fa';
 import { FaUserPlus } from 'react-icons/fa';
 import { useNavigate } from 'react-router';
 import { Link as RouterLink } from 'react-router-dom';
-import { IconAvatar, Label, useCardStyle } from '../../../../design';
+import { IconAvatar, Label, useCardStyle } from 'design';
 
-const OrganizationListItem = (props) => {
+export const OrganizationListItem = (props) => {
   const { organization } = props;
   const classes = useCardStyle();
   const navigate = useNavigate();
@@ -241,4 +241,3 @@ const OrganizationListItem = (props) => {
 OrganizationListItem.propTypes = {
   organization: PropTypes.object.isRequired
 };
-export default OrganizationListItem;
