+++ frontend/src/modules/Organizations/services/removeGroup.js
