+++ frontend/src/modules/Organizations/views/OrganizationEditForm.js
@@ -25,13 +25,13 @@ import {
   ChevronRightIcon,
   ChipInput,
   useSettings
-} from '../../../../design';
-import { SET_ERROR, useDispatch } from '../../../../globalErrors';
+} from 'design';
+import { SET_ERROR, useDispatch } from 'globalErrors';
+import { useClient } from 'services';
 import {
   getOrganization,
-  updateOrganization,
-  useClient
-} from '../../../../services';
+  updateOrganization
+} from '../services';
 
 const OrganizationEditForm = (props) => {
   const navigate = useNavigate();
