--- frontend/src/services/graphql/Organization/index.js
@@ -1,19 +1,0 @@
-export * from './addOrUpdateOrganizationTopic';
-export * from './archiveOrganization';
-export * from './createGroup';
-export * from './createOrganization';
-export * from './getOrganization';
-export * from './inviteGroup';
-export * from './inviteUser';
-export * from './listInvitedGroups';
-export * from './listOrganizationEnvironmentGroups';
-export * from './listOrganizationGroups';
-export * from './listOrganizationTopics';
-export * from './listOrganizationUsers';
-export * from './listOrganizations';
-export * from './removeGroup';
-export * from './removeUser';
-export * from './updateGroupRoleInOrganization';
-export * from './updateOrganization';
-export * from './updateTopic';
-export * from './updateUserRole';
