--- frontend/src/services/graphql/Organization/inviteUser.js
@@ -1,16 +1,0 @@
-import { gql } from 'apollo-boost';
-
-export const inviteUser = ({ organizationUri, userName, role }) => ({
-  variables: {
-    input: { organizationUri, userName, role: role || 'Member' }
-  },
-  mutation: gql`
-    mutation InviteUser($input: NewOrganizationUserInput) {
-      inviteUser(input: $input) {
-        userName
-        userRoleInOrganization
-        created
-      }
-    }
-  `
-});
