--- frontend/src/services/graphql/Organization/removeUser.js
@@ -1,12 +1,0 @@
-import { gql } from 'apollo-boost';
-
-export const removeUser = ({ input }) => ({
-  variables: {
-    input
-  },
-  mutation: gql`
-    mutation RemoveUser($input: RemoveOrganizationUserInput) {
-      removeUser(input: $input)
-    }
-  `
-});
