--- frontend/src/services/graphql/Organization/updateGroupRoleInOrganization.js
@@ -1,19 +1,0 @@
-import { gql } from 'apollo-boost';
-
-export const updateGroupRoleInOrganization = ({ groupUri, role }) => ({
-  variables: {
-    input: { role: role || 'Member' },
-    groupUri
-  },
-  mutation: gql`
-    mutation UpdateGroup($groupUri: String, $input: ModifyGroupInput) {
-      updateGroup(groupUri: $groupUri, input: $input) {
-        groupUri
-        groupRoleInOrganization
-        userRoleInGroup
-        created
-        updated
-      }
-    }
-  `
-});
