--- frontend/src/services/graphql/Organization/updateUserRole.js
@@ -1,16 +1,0 @@
-import { gql } from 'apollo-boost';
-
-export const updateUserRole = ({ organizationUri, userName, role }) => ({
-  variables: {
-    input: { organizationUri, userName, role: role || 'Member' }
-  },
-  mutation: gql`
-    mutation UpdateUser($input: ModifyOrganizationUserInput) {
-      updateUser(input: $input) {
-        userName
-        userRoleInOrganization
-        created
-      }
-    }
-  `
-});
